where would we be without a state variabe filter
hit a wall with the envelope thing. they should be a ratio of which frequency?! how do i know the length of the gate beforehand. knowing the frequency of each lfo (lol) alone won't help, i need to figure out how they interfere based on each gate's frequency and width. and like i don't even know if it's gonna be that interesting if i do make it.. eh too lazy